requirements.txt
Django==3.0.5
django-cors-headers==3.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djongo==1.3.3
dnspython==2.1.0
instamojo-wrapper==1.2.0
PyJWT==1.7.1
pymongo==3.11.2
pytz==2020.5
requests==2.25.1
sqlparse==0.2.4
urllib3==1.26.2
whitenoise==5.2.0
